Make hostname handling more flexible and add metrics.
Since there are multiple ways to specify a hostname to reboot, this
change aims to handle to most common ones, i.e. the same we already
accept in other tools (such as drac.py).

reboot_drac_duration_seconds is a histogram containing the time it
took to reboot a DRAC. Bucket size has been chosen empirically and can
probably be improved.