Add SSH connection timeout.
This should be significantly longer than the average time it takes to
connect to a DRAC and process the reboot command, and shorter than
the timeout set in clients (e.g. Rebot has a 2 minutes timeout).
The previous configuration - with no timeouts - apparently kept
retrying the connection more than once, which is less than ideal.